requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
environs==8.0.0
frozenlist==1.3.3
idna==3.4
instaloader==4.9.6
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
packaging==23.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
urllib3==1.26.14
yarl==1.8.2
